+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `SuperficialValidation.validateType(...)` and `SuperficialValidation.validateTypes(...)` helper methods.
+
 ### [v0.12](https://github.com/realityforge/proton/tree/v0.12) (2020-01-09) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.11...v0.12)
 
 Changes in this release: